+++ x/feemarket/genesis.go
@@ -20,15 +20,12 @@ func InitGenesis(
 		panic(errorsmod.Wrap(err, "could not set parameters at genesis"))
 	}
 
-	k.SetBlockGasUsed(ctx, data.BlockGas)
-
 	return []abci.ValidatorUpdate{}
 }
 
 // ExportGenesis exports genesis state of the fee market module
 func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 	return &types.GenesisState{
-		Params:   k.GetParams(ctx),
-		BlockGas: k.GetBlockGasUsed(ctx),
+		Params: k.GetParams(ctx),
 	}
 }