+++ x/feemarket/keeper/abci.go
@@ -1,8 +1,6 @@
 package keeper
 
 import (
-	"fmt"
-
 	"github.com/EscanBE/evermint/v12/x/feemarket/types"
 	abci "github.com/cometbft/cometbft/abci/types"
 
@@ -10,50 +8,41 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// BeginBlock updates base fee
-func (k *Keeper) BeginBlock(ctx sdk.Context, _ abci.RequestBeginBlock) {
-	baseFee := k.CalculateBaseFee(ctx)
+// EndBlock update base fee for the next block.
+// The EVM end block logic doesn't update the validator set, thus it returns an empty slice.
+func (k Keeper) EndBlock(ctx sdk.Context, _ abci.RequestEndBlock) {
+	k.updateBaseFeeForNextBlock(ctx)
+}
 
-	// return immediately if base fee is nil
-	if baseFee == nil {
+func (k Keeper) updateBaseFeeForNextBlock(ctx sdk.Context) {
+	if ctx.BlockGasMeter() == nil {
+		k.Logger(ctx).Error("block gas meter is nil when setting base fee for the next block")
 		return
 	}
+
+	baseFee := k.CalculateBaseFee(ctx)
 
 	k.SetBaseFee(ctx, baseFee)
 
 	defer func() {
-		telemetry.SetGauge(float32(baseFee.Int64()), "feemarket", "base_fee")
+		telemetry.SetGauge(func() float32 {
+			if baseFee == nil {
+				return 0.0
+			}
+			return float32(baseFee.Int64())
+		}(), "feemarket", "base_fee")
 	}()
 
-	// Store current base fee in event
+	// Store next base fee in event
 	ctx.EventManager().EmitEvents(sdk.Events{
 		sdk.NewEvent(
 			types.EventTypeFeeMarket,
-			sdk.NewAttribute(types.AttributeKeyBaseFee, baseFee.String()),
+			sdk.NewAttribute(types.AttributeKeyBaseFee, func() string {
+				if baseFee == nil {
+					return "0"
+				}
+				return baseFee.String()
+			}()),
 		),
 	})
-}
-
-// EndBlock update block gas used.
-// The EVM end block logic doesn't update the validator set, thus it returns
-// an empty slice.
-func (k *Keeper) EndBlock(ctx sdk.Context, _ abci.RequestEndBlock) {
-	if ctx.BlockGasMeter() == nil {
-		k.Logger(ctx).Error("block gas meter is nil when setting block gas used")
-		return
-	}
-
-	gasUsed := ctx.BlockGasMeter().GasConsumedToLimit()
-
-	k.SetBlockGasUsed(ctx, gasUsed)
-
-	defer func() {
-		telemetry.SetGauge(float32(gasUsed), "feemarket", "block_gas")
-	}()
-
-	ctx.EventManager().EmitEvent(sdk.NewEvent(
-		"block_gas",
-		sdk.NewAttribute("height", fmt.Sprintf("%d", ctx.BlockHeight())),
-		sdk.NewAttribute("amount", fmt.Sprintf("%d", gasUsed)),
-	))
 }