+++ x/feemarket/keeper/grpc_query.go
@@ -3,7 +3,6 @@ package keeper
 import (
 	"context"
 
-	errorsmod "cosmossdk.io/errors"
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -35,18 +34,4 @@ func (k Keeper) BaseFee(c context.Context, _ *types.QueryBaseFeeRequest) (*types
 	}
 
 	return res, nil
-}
-
-// BlockGas implements the Query/BlockGas gRPC method
-func (k Keeper) BlockGas(c context.Context, _ *types.QueryBlockGasRequest) (*types.QueryBlockGasResponse, error) {
-	ctx := sdk.UnwrapSDKContext(c)
-	gas := sdkmath.NewIntFromUint64(k.GetBlockGasUsed(ctx))
-
-	if !gas.IsInt64() {
-		return nil, errorsmod.Wrapf(sdk.ErrIntOverflowCoin, "block gas %s is higher than MaxInt64", gas)
-	}
-
-	return &types.QueryBlockGasResponse{
-		Gas: gas.Int64(),
-	}, nil
 }